+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.1.6] - 2020-05-22
+### Changed
+* core/src/main/resources/Samples/Input_Json_Specification.json
+* core/src/main/scala/core/Controller.scala
+* core/src/main/scala/core/DataFrameFromTo.scala
+* core/src/main/scala/core/DataPull.scala
+* core/src/main/scala/core/Migration.scala
+* core/src/main/scala/logging/DataPullLog.scala
+
 ## [0.1.5] - 2020-05-21
 ### Changed
 * CHANGELOG.md
@@ -11,7 +20,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * core/src/main/resources/Samples/Input_Json_Specification.json
 * core/src/main/scala/core/DataFrameFromTo.scala
 * core/src/main/scala/core/Migration.scala
-
 
 ## [0.1.4] - 2020-05-15
 ### Changed